refactor: move canonical-ensemble models to HRGCanonical module
Move ThermalModelCanonical, ThermalModelCanonicalCharm, and
ThermalModelCanonicalStrangeness from HRGBase into a new HRGCanonical
module, matching how the other model families are organized (HRGEV,
HRGVDW, HRGRealGas, HRGPCE, ...). HRGBase is now limited to base
infrastructure (particles, base/ideal models, math). The dependency is
one-way: HRGCanonical -> HRGBase.

Backward compatibility is preserved:
- Forwarding shim headers remain at the old include/HRGBase/ paths and
  redirect to the HRGCanonical/ headers.
- The HRGBase.h umbrella still pulls in the canonical headers.
- A new HRGCanonical.h umbrella is added for new code.

No behavior change: class names and the thermalfist namespace are
unchanged, only header include paths move. Internal consumers (HRGEV,
HRGVDW, HRGEventGenerator, GUI) are updated to the new paths.
