+++ include/HRGBase.h
@@ -10,9 +10,9 @@
 #include "HRGBase/SplineFunction.h"
 #include "HRGBase/ThermalModelIdeal.h"
 #include "HRGBase/ThermalModelBase.h"
-#include "HRGBase/ThermalModelCanonical.h"
-#include "HRGBase/ThermalModelCanonicalCharm.h"
-#include "HRGBase/ThermalModelCanonicalStrangeness.h"
+#include "HRGCanonical/ThermalModelCanonical.h"
+#include "HRGCanonical/ThermalModelCanonicalCharm.h"
+#include "HRGCanonical/ThermalModelCanonicalStrangeness.h"
 #include "HRGBase/ThermalParticle.h"
 #include "HRGBase/ThermalParticleSystem.h"
 #include "HRGBase/Utility.h"
