+++ include/HRGCanonical.h
@@ -1,0 +1,10 @@
+/*
+ * Thermal-FIST package
+ *
+ * Copyright (c) 2015-2025 Volodymyr Vovchenko
+ *
+ * GNU General Public License (GPLv3 or later)
+ */
+#include "HRGCanonical/ThermalModelCanonical.h"
+#include "HRGCanonical/ThermalModelCanonicalCharm.h"
+#include "HRGCanonical/ThermalModelCanonicalStrangeness.h"
