+++ src/gui/QtThermalFIST/eventgeneratortab.cpp
@@ -29,9 +29,9 @@
 #include "HRGEV/ThermalModelEVCrossterms.h"
 #include "HRGEV/ExcludedVolumeHelper.h"
 #include "HRGVDW/ThermalModelVDW.h"
-#include "HRGBase/ThermalModelCanonical.h"
-#include "HRGBase/ThermalModelCanonicalStrangeness.h"
-#include "HRGBase/ThermalModelCanonicalCharm.h"
+#include "HRGCanonical/ThermalModelCanonical.h"
+#include "HRGCanonical/ThermalModelCanonicalStrangeness.h"
+#include "HRGCanonical/ThermalModelCanonicalCharm.h"
 #include "HRGEventGenerator/SphericalBlastWaveEventGenerator.h"
 #include "HRGEventGenerator/CylindricalBlastWaveEventGenerator.h"
 #include "HRGEventGenerator/CracowFreezeoutEventGenerator.h"
