+++ src/library/HRGCanonical/ThermalModelCanonicalCharm.cpp
@@ -5,7 +5,7 @@
  *
  * GNU General Public License (GPLv3 or later)
  */
-#include "HRGBase/ThermalModelCanonicalCharm.h"
+#include "HRGCanonical/ThermalModelCanonicalCharm.h"
 
 #include <cassert>
 #include <stdexcept>
