+++ src/library/HRGEventGenerator/EventGeneratorBase.cpp
@@ -14,8 +14,8 @@
 #include "HRGBase/xMath.h"
 #include "HRGBase/ThermalModelBase.h"
 #include "HRGBase/ThermalModelIdeal.h"
-#include "HRGBase/ThermalModelCanonicalStrangeness.h"
-#include "HRGBase/ThermalModelCanonical.h"
+#include "HRGCanonical/ThermalModelCanonicalStrangeness.h"
+#include "HRGCanonical/ThermalModelCanonical.h"
 #include "HRGEV/ThermalModelEVDiagonal.h"
 #include "HRGEV/ThermalModelEVCrossterms.h"
 #include "HRGEV/ExcludedVolumeHelper.h"
